I kinda redid the points for touching the ladder, and also used arm since before we werent using arm
(All changes were made in the red_neg_awp_redo funtion.)

I decieded to use the arm since it like makes sure that we touch ladder, and lets me take out all the ugly zipties :)

Also redid the MoveToPose for the ladder bc it was going to far forward and with the arm it will kill the arm from impact. So I moved the point more back.